refactor(js): extract Google Docs fallback into gdocs-fallback.js
Moves the browser-to-export fallback helpers out of gdocs.js to keep the
module under the 1500-line lint limit after merging main (issue #92 R5/R6
additions). Public API is unchanged.
