+++ js/src/gdocs.js
@@ -13,8 +13,18 @@ import { convertHtmlToMarkdown } from './lib.js';
 import { createBrowser as defaultCreateBrowser } from './browser.js';
 import { preprocessGoogleDocsExportHtml } from './gdocs-preprocess.js';
 import { localizeGoogleDocsModelImages } from './gdocs-images.js';
+import {
+  googleDocsBrowserModelUnavailableError,
+  isGoogleDocsBrowserModelUnavailableError,
+  fetchGoogleDocByExportFormat as fetchGoogleDocByExportFormatImpl,
+  captureGoogleDocWithBrowserOrFallback as captureGoogleDocWithBrowserOrFallbackImpl,
+} from './gdocs-fallback.js';
 
-export { preprocessGoogleDocsExportHtml, localizeGoogleDocsModelImages };
+export {
+  preprocessGoogleDocsExportHtml,
+  localizeGoogleDocsModelImages,
+  isGoogleDocsBrowserModelUnavailableError,
+};
 
 const GDOCS_URL_PATTERN = /docs\.google\.com\/document\/d\/([a-zA-Z0-9_-]+)/;
 
@@ -22,7 +32,6 @@ const GDOCS_EXPORT_BASE = 'https://docs.google.com/document/d';
 const GDOCS_API_BASE = 'https://docs.googleapis.com/v1/documents';
 const DEFAULT_USER_AGENT =
   'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/120.0.0.0 Safari/537.36';
-const GDOCS_BROWSER_MODEL_UNAVAILABLE = 'GDOCS_BROWSER_MODEL_UNAVAILABLE';
 
 /**
  * Supported Google Docs export formats.
@@ -311,119 +320,39 @@ export async function fetchGoogleDocAsMarkdown(url, options = {}) {
 
 /**
  * Fetch a Google Doc through the public export pipeline for a requested output
- * format. This is used as the lossless fallback when browser-model capture
- * cannot read DOCS_modelChunk data from the editor page.
+ * format. Thin wrapper over the fallback module that injects sibling fetchers.
  *
  * @param {string} url - Google Docs URL
  * @param {Object} [options] - Options
- * @param {string} [options.format='markdown'] - Requested CLI output format
- * @param {string} [options.apiToken] - API token for private documents
- * @param {Object} [options.log] - Optional logger
  * @returns {Promise<Object>} Export result normalized for CLI rendering
  */
-export async function fetchGoogleDocByExportFormat(url, options = {}) {
-  const { format = 'markdown', apiToken, log } = options;
-  const normalized = (format || 'markdown').toLowerCase();
-
-  if (normalized === 'archive' || normalized === 'zip') {
-    const archiveResult = await fetchGoogleDocAsArchive(url, { apiToken, log });
-    return {
-      ...archiveResult,
-      content: archiveResult.markdown,
-      sourceFormat: 'archive',
-    };
-  }
-
-  if (normalized === 'markdown' || normalized === 'md') {
-    const result = await fetchGoogleDocAsMarkdown(url, { apiToken, log });
-    return {
-      ...result,
-      content: result.markdown,
-      sourceFormat: 'markdown',
-    };
-  }
-
-  if (normalized === 'html') {
-    const result = await fetchGoogleDoc(url, {
-      format: 'html',
-      apiToken,
-      log,
-    });
-    return {
-      ...result,
-      html: result.content,
-      sourceFormat: 'html',
-    };
-  }
-
-  if (normalized === 'txt' || normalized === 'text') {
-    const result = await fetchGoogleDoc(url, {
-      format: 'txt',
-      apiToken,
-      log,
-    });
-    return {
-      ...result,
-      text: result.content,
-      sourceFormat: 'txt',
-    };
-  }
-
-  throw new Error(
-    `Unsupported Google Docs export fallback format "${format}".`
+export function fetchGoogleDocByExportFormat(url, options = {}) {
+  return fetchGoogleDocByExportFormatImpl(
+    { fetchGoogleDoc, fetchGoogleDocAsMarkdown, fetchGoogleDocAsArchive },
+    url,
+    options
   );
 }
 
 /**
  * Capture a Google Doc through the browser model, falling back to public export
- * when the editor does not expose model chunks.
+ * when the editor does not expose model chunks. Thin wrapper.
  *
  * @param {string} url - Google Docs URL
  * @param {Object} [options] - Capture options
- * @param {string} [options.format='markdown'] - Requested output format
- * @param {boolean} [options.fallback=true] - Whether to use export fallback
- * @param {Function} [options.onFallback] - Called with browser error
  * @returns {Promise<Object>} Browser-model or public-export result
  */
-export async function captureGoogleDocWithBrowserOrFallback(url, options = {}) {
-  const {
-    format = 'markdown',
-    fallback = true,
-    onFallback,
-    ...browserOptions
-  } = options;
-
-  try {
-    const result = await captureGoogleDocWithBrowser(url, browserOptions);
-    return {
-      ...result,
-      method: 'browser-model',
-      fallback: false,
-    };
-  } catch (err) {
-    if (!fallback || !isGoogleDocsBrowserModelUnavailableError(err)) {
-      throw err;
-    }
-
-    onFallback?.(err);
-    browserOptions.log?.warn?.(() => ({
-      event: 'gdocs.browser-model.fallback-public-export',
-      reason: err.message,
-      format,
-    }));
-
-    const result = await fetchGoogleDocByExportFormat(url, {
-      format,
-      apiToken: browserOptions.apiToken,
-      log: browserOptions.log,
-    });
-    return {
-      ...result,
-      method: 'public-export',
-      fallback: true,
-      browserError: err.message,
-    };
-  }
+export function captureGoogleDocWithBrowserOrFallback(url, options = {}) {
+  return captureGoogleDocWithBrowserOrFallbackImpl(
+    {
+      captureGoogleDocWithBrowser,
+      fetchGoogleDoc,
+      fetchGoogleDocAsMarkdown,
+      fetchGoogleDocAsArchive,
+    },
+    url,
+    options
+  );
 }
 
 /**
@@ -545,25 +474,6 @@ export async function captureGoogleDocWithBrowser(url, options = {}) {
     }
     await browser.close();
   }
-}
-
-/**
- * Check whether an error means Google Docs browser-model data was unavailable.
- *
- * @param {Error} err - Error to classify
- * @returns {boolean} True if export fallback is appropriate
- */
-export function isGoogleDocsBrowserModelUnavailableError(err) {
-  return (
-    err?.code === GDOCS_BROWSER_MODEL_UNAVAILABLE ||
-    String(err?.message || '').includes('did not expose DOCS_modelChunk data')
-  );
-}
-
-function googleDocsBrowserModelUnavailableError(message) {
-  const err = new Error(message);
-  err.code = GDOCS_BROWSER_MODEL_UNAVAILABLE;
-  return err;
 }
 
 async function installDocsModelCapture(page) {
